% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_pick_winner_shiny.R
\name{bayes.pick.winner}
\alias{bayes.pick.winner}
\title{Bayesian pick the winner design}
\usage{
bayes.pick.winner()
}
\description{
This function will run Shiny application for the Bayesian pick the winner design in randomized phase II trials.
             two-arm stage one and two and a two stage mulitple arm design.
}
